Control.Monad.ListT.Funcs

Description

List functions with type limited to use ListT. This might come useful for type interference.

Functions where the List is an input type and not only the result type do not need special limited versions.

Documentation

iterateM :: Monad m => (a -> m a) -> m a -> ListT m a

repeatM :: Monad m => m a -> ListT m a

repeat :: Monad m => a -> ListT m a

fromList :: Monad m => [a] -> ListT m a